feat: products authorizations!

+++ resources/views/pages/products/⚡delete.blade.php
@@ -13,28 +13,37 @@ new class extends Component
     }
 
     public function delete(){
+        Gate::authorize("update",$this->product);
         // O usuário aceitou, deleta
         $this->product->delete();
 
         // Mensagem bonitinha
-        redirect()->route('products')->with('delete','Deletado com sucesso!');
+        redirect()->route('products.products')->with('delete','Deletado com sucesso!');
     }
 };
 ?>
 
-<div class="modal show" style="display: block;" tabindex="-1">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title">Apagar {{ $product->name }}...</h5>
-      </div>
-      <div class="modal-body">
-        <p>Tem certeza que quer apagar {{ $product->name }}?</p>
-      </div>
-      <div class="modal-footer">
-        <a href="{{ route('products') }}" class="btn btn-primary">Close</a>
-        <button type="button" wire:click='delete' class="btn btn-secondary">Delete</button>
+
+
+@can('delete',$product)
+  <div class="modal show" style="display: block;" tabindex="-1">
+    <div class="modal-dialog">
+      <div class="modal-content">
+        <div class="modal-header">
+          <h5 class="modal-title">Apagar {{ $product->name }}...</h5>
+        </div>
+        <div class="modal-body">
+          <p>Tem certeza que quer apagar {{ $product->name }}?</p>
+        </div>
+        <div class="modal-footer">
+          <a href="{{ route('products.products') }}" class="btn btn-primary">Close</a>
+          <button type="button" wire:click='delete' class="btn btn-secondary">Delete</button>
+        </div>
       </div>
     </div>
   </div>
+@else
+<div class="container p-3 m-3">
+  <h4>Você não possui este produto!</h4>
 </div>
+@endcan
